refactor(app): reorganize module imports, remove unused queues, and update controller paths for improved structure and maintainability
