+++ src/app.module.ts
@@ -6,26 +6,21 @@ import { CacheModule } from '@nestjs/cache-manager';
 import { Module } from '@nestjs/common';
 import { ConfigModule } from '@nestjs/config';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { AccountModule } from './plugins/account/account.module';
 import { AePricingModule } from './ae-pricing/ae-pricing.module';
 import { AeModule } from './ae/ae.module';
-import { AffiliationModule } from './plugins/affiliation/affiliation.module';
 import { AnalyticsModule } from './analytics/analytics.module';
 import { AppController } from './app.controller';
 import { AppService } from './app.service';
-import { BclModule } from './plugins/bcl/bcl.module';
 import { DATABASE_CONFIG, REDIS_CONFIG } from './configs';
 import { MdwModule } from './mdw-sync/mdw.module';
+import { AccountModule } from './plugins/account/account.module';
+import { AffiliationModule } from './plugins/affiliation/affiliation.module';
+import { BclModule } from './plugins/bcl/bcl.module';
+import { TokensModule } from './plugins/bcl/tokens.module';
+import { TransactionsModule } from './plugins/bcl/transactions.module';
 import { DexPluginModule } from './plugins/dex/dex-plugin.module';
 import { SocialPluginModule } from './plugins/social/social-plugin.module';
 import { TippingPluginModule } from './plugins/tipping/tipping-plugin.module';
-import {
-  DELETE_OLD_TOKENS_QUEUE,
-  PULL_TOKEN_INFO_QUEUE,
-  SYNC_TOKEN_HOLDERS_QUEUE,
-} from './tokens/queues/constants';
-import { TokensModule } from './plugins/bcl/tokens.module';
-import { TransactionsModule } from './plugins/bcl/transactions.module';
 import { TrendingTagsModule } from './trending-tags/trending-tags.module';
 
 @Module({
@@ -38,17 +33,6 @@ import { TrendingTagsModule } from './trending-tags/trending-tags.module';
     BullModule.forRoot({
       redis: REDIS_CONFIG,
     }),
-    BullModule.registerQueue(
-      {
-        name: PULL_TOKEN_INFO_QUEUE,
-      },
-      {
-        name: SYNC_TOKEN_HOLDERS_QUEUE,
-      },
-      {
-        name: DELETE_OLD_TOKENS_QUEUE,
-      },
-    ),
     TypeOrmModule.forRoot({
       ...DATABASE_CONFIG,
       entities: [__dirname + '/**/entities/*.entity{.ts,.js}'],
