+++ src/app.service.ts
@@ -1,12 +1,9 @@
-import { InjectQueue } from '@nestjs/bull';
 import { Injectable } from '@nestjs/common';
 import { Cron, CronExpression } from '@nestjs/schedule';
-import { Queue } from 'bull';
 import moment, { Moment } from 'moment';
 import { AePricingService } from './ae-pricing/ae-pricing.service';
 import { CommunityFactoryService } from './ae/community-factory.service';
 import { WebSocketService } from './ae/websocket.service';
-import { DELETE_OLD_TOKENS_QUEUE } from './tokens/queues/constants';
 @Injectable()
 export class AppService {
   startedAt: Moment;
@@ -14,9 +11,6 @@ export class AppService {
     private communityFactoryService: CommunityFactoryService,
     private aePricingService: AePricingService,
     private websocketService: WebSocketService,
-
-    @InjectQueue(DELETE_OLD_TOKENS_QUEUE)
-    private readonly deleteOldTokensQueue: Queue,
   ) {
     this.init();
     this.startedAt = moment();
@@ -25,12 +19,6 @@ export class AppService {
 
   async init() {
     await this.aePricingService.pullAndSaveCoinCurrencyRates();
-
-    const factory = await this.communityFactoryService.getCurrentFactory();
-    await this.deleteOldTokensQueue.empty();
-    void this.deleteOldTokensQueue.add({
-      factories: [factory.address],
-    });
   }
 
   setupLiveSync() {
